requirements.txt
# Core
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-environ>=0.12,<1.0

# Database
psycopg2-binary>=2.9,<3.0

# Task Queue
celery>=5.4,<6.0
django-celery-beat>=2.7,<3.0
redis>=5.2,<6.0

# BTC Price
requests>=2.32,<3.0

# CORS (for frontend API calls from templates)
django-cors-headers>=4.6,<5.0

# Production server
gunicorn>=23.0,<24.0
